+++ src/review0/main.py
@@ -10,20 +10,17 @@ app = typer.Typer()
 console = Console()
 
 @app.command()
-def main(content_file: Annotated[str, typer.Argument(help="Path to file containing the content to be reviewed")], criteria_file: Annotated[str, typer.Argument(help="Path to file containing the review criteria")], model_name: str = "mistral", temperature: float = 0.7):
-    """Console script for review0."""
+def review(content_file: Annotated[str, typer.Argument(help="Path to file containing the content to be reviewed")], criteria_file: Annotated[str, typer.Argument(help="Path to file containing the review criteria")], model_name: str = "mistral", temperature: float = 0.7, num_ctx: int = 4096):
+    """Console script for review0 technical review functionality."""
     console.print("Review0: The baseline review tool")
-    #model_name = "mistral" # "gemma3" "openchat"
-    #temperature = 0.8 # 1 #0.1
-    #content_file = "../../data/test2.pdf"
     # Parse PDF
     text, text_highlight = pdf_to_text(content_file, [("Abstract", "Introduction")]) 
     #print(text, "\n----------------\n")
-    print(text_highlight, "\n----------------\n")
+    #print(text_highlight, "\n----------------\n")
     # Get summary from the highlight using LLM
     if text_highlight != "":
         prompt = "Generate bullet point technical summary and assess scientific and business merit for the following proposal. Include main the challenges, strengths, weaknesses and potential risks."
-        model = load_model(model_name, temperature, num_predict = 1024, num_ctx=4096) 
+        model = load_model(model_name, temperature, num_predict = 1024, num_ctx=num_ctx) 
         result = query(model, prompt + text_highlight)
     print(result.content)
     print("\n----------------\n\n")
@@ -31,7 +28,7 @@ def main(content_file: Annotated[str, typer.Argument(help="Path to file containi
     #criteria_file = "../../data/criteria.txt"
     questions = get_criteria(criteria_file)
     # Get answers to questions using RAG
-    vector_store = load_vectorstore("nomic-embed-text", [content_file], [text]) 
+    vector_store, _ = load_vectorstore("nomic-embed-text", [content_file], [text]) 
     for question in questions:
         print("Question:", question)
         prompt = get_qarag_prompt(model_name, vector_store, "What is the " + question) 
@@ -40,6 +37,23 @@ def main(content_file: Annotated[str, typer.Argument(help="Path to file containi
         print("\n----------------\n")
     return
 
+@app.command()
+def revise(content_file: Annotated[str, typer.Argument(help="Path to file containing the content to be revised for language/grammer")], model_name: str = "mistral", temperature: float = 1, num_ctx: int = 4096):
+    """Console script for review0 language revision functionality."""
+    console.print("Review0: The baseline review tool")
+    # Parse PDF
+    text, _ = pdf_to_text(content_file) 
+    # Get split texts from the document 
+    _, all_splits = load_vectorstore("nomic-embed-text", [content_file], [text], chunk_size=4000, chunk_overlap=40) 
+    #print(len(all_splits))
+    # Iterate over splits and check language issues
+    prompt = "Highlight major grammer issues and spelling errors in the following text and suggest corrections as a bullet list. Ignore whitespace, punctuation and capitilization errors." # any language issues
+    model = load_model(model_name, temperature, num_predict = 512, num_ctx=num_ctx) 
+    for text_split in all_splits:
+        result = query(model, prompt + text_split.page_content)
+        print(result.content)
+        print("\n----------------\n\n")
+    return
 
 if __name__ == "__main__":
     app()